GPG-429 Move static pages (Contact Us, etc.) to new StaticPagesController (not based on BaseController)
